Extracted out parser to handle commands
