+++ src/main/java/Duke.java
@@ -1,96 +1,10 @@
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Scanner;
 
 
 public class Duke {
-    private static Scanner sc = new Scanner(System.in);
     private static ArrayList<Task> taskList = new ArrayList<>();
     private static String DATA_FILE_PATH = "src/main/java/data.txt";
     private static Storage storage = new Storage(DATA_FILE_PATH);
-
-    private static void lineSpacing() {
-        System.out.println("____________________________________________________________");
-    }
-
-    private static void displayAddedTask(Task currentTask) {
-        lineSpacing();
-        System.out.println("Got it. I've added this task: ");
-        System.out.println(currentTask);
-        System.out.println(String.format("Now you have %d tasks in the list.", taskList.size()));
-        lineSpacing();
-    }
-
-    private static void printList() {
-        lineSpacing();
-        for (int i = 0; i < taskList.size(); i++) {
-            Task currentTask = taskList.get(i);
-            System.out.println(String.format("%d.%s", i + 1, currentTask));
-        }
-        lineSpacing();
-    }
-
-    private static void doneTask(String userInput) {
-        String[] inputArray = userInput.split(" ");
-        Task completedTask = taskList.get(Integer.parseInt(inputArray[1]) - 1);
-        completedTask.markAsDone();
-        lineSpacing();
-        System.out.println("Nice! I've marked this task as done:");
-        System.out.println(completedTask);
-        lineSpacing();
-        storage.saveData(taskList);
-    }
-
-    private static void addDeadline(String userInput) {
-        List<String> inputArray = Arrays.asList(userInput.split(" /by "));
-        String by = inputArray.get(1);
-        ArrayList<String> descriptionArray = new ArrayList<String>(Arrays.asList(inputArray.get(0).split(" ")));
-        descriptionArray.remove(0);
-        String description = String.join(" ", descriptionArray);
-        Deadline newDeadline = new Deadline(description, by);
-        taskList.add(newDeadline);
-        displayAddedTask(newDeadline);
-        storage.saveData(taskList);
-    }
-
-    private static void addEvent(String userInput) {
-        List<String> inputArray = Arrays.asList(userInput.split(" /at "));
-        String timeFrame = inputArray.get(1);
-        ArrayList<String> descriptionArray = new ArrayList<String>(Arrays.asList(inputArray.get(0).split(" ")));
-        descriptionArray.remove(0);
-        String description = String.join(" ", descriptionArray);
-        Event newEvent = new Event(description, timeFrame);
-        taskList.add(newEvent);
-        displayAddedTask(newEvent);
-        storage.saveData(taskList);
-    }
-
-    private static void addTodo(String userInput) throws DukeException {
-        List<String>  inputArray = Arrays.asList(userInput.split(" "));
-        if (inputArray.size() <= 1) {
-            throw new DukeException("todo");
-        }
-        ArrayList<String> descriptionArray =  new ArrayList<String>(inputArray);
-        descriptionArray.remove(0);
-        String description = String.join(" ",descriptionArray);
-        Todo newTodo = new Todo(description);
-        taskList.add(newTodo);
-        displayAddedTask(newTodo);
-        storage.saveData(taskList);
-    }
-
-    private static void deleteTask(String userInput) {
-        String[] inputArray = userInput.split(" ");
-        Task removedTask = taskList.remove(Integer.parseInt(inputArray[1]) - 1);
-        lineSpacing();
-        System.out.println("Noted. I've removed this task: ");
-        System.out.println(removedTask);
-        System.out.println(String.format("Now you have %d tasks in the list.", taskList.size()));
-        lineSpacing();
-        storage.saveData(taskList);
-    }
-
 
 
     public static void main(String[] args) {
@@ -101,53 +15,10 @@ public class Duke {
                 + "|____/ \\__,_|_|\\_\\___|\n";
         System.out.println("Hello from\n" + logo);
         taskList = storage.loadData();
-        while (true) {
-            try {
-                String userInput = sc.nextLine();
-                if (userInput.equals("bye")) {
-                    lineSpacing();
-                    System.out.println("Bye. Hope to see you again soon!");
-                    lineSpacing();
-                    break;
-                }
-
-                if (userInput.equals("list")) {
-                    printList();
-                    continue;
-                }
-
-                if (userInput.startsWith("done")) {
-                    doneTask(userInput);
-                    continue;
-                }
-
-                if (userInput.startsWith("deadline")) {
-                    addDeadline(userInput);
-                    continue;
-                }
-
-                if (userInput.startsWith("event")) {
-                    addEvent(userInput);
-                    continue;
-                }
-
-                if (userInput.startsWith(("todo"))) {
-                    addTodo(userInput);
-                    continue;
-                }
-
-                if (userInput.startsWith("delete")) {
-                    deleteTask(userInput);
-                    continue;
-                }
-
-                throw new DukeException("unknown command");
-            } catch (DukeException e) {
-                lineSpacing();
-                System.out.println(e);
-                lineSpacing();
-            }
+        Parser parser = new Parser(storage, taskList);
+        parser.parseCommands();
         }
-    }
 
 }
+
+
